+++ caseworker/advice/services.py
@@ -350,6 +350,14 @@ def unassessed_trigger_list_goods(case):
     ]
 
 
+def assessed_trigger_list_goods(case):
+    return [
+        product
+        for product in filter_trigger_list_products(case["data"]["goods"])
+        if self.is_trigger_list_assessed(product)
+    ]
+
+
 def get_advice_tab_context(case, caseworker, queue_id):
     """Get contextual information for the advice tab such as the tab's URL and
     button visibility, based off the case, the current user and current user's queue.
@@ -393,7 +401,7 @@ def get_advice_tab_context(case, caseworker, queue_id):
                 context["buttons"]["move_case_forward"] = True
 
             # BEIS Nuclear need to assess products first before giving recommendation
-            if team_alias == BEIS_NUCLEAR and queue_alias == BEIS_NUCLEAR_CASES_TO_REVIEW and not existing_advice:
+            if team_alias in BEIS_NUCLEAR and queue_alias == BEIS_NUCLEAR_CASES_TO_REVIEW and not existing_advice:
                 context["buttons"]["make_recommendation"] = len(unassessed_trigger_list_goods(case)) == 0
                 context["buttons"]["assess_trigger_list_products"] = len(unassessed_trigger_list_goods(case)) > 0
